fix(approvals): address ClawSweeper review (qqbot per-account, hot-path split)
Local ClawSweeper review of the head surfaced issues the first autoreview
missed:

- QQBot markdown is per-account (markdownSupport); false accounts send raw
  msg_type 0 text. A channel-wide "markdown" declaration would leave literal
  fences on those accounts, so revert QQBot to the plaintext default — the only
  safe channel-wide choice for a static field. (This reverts a declaration
  added in error responding to the prior review.)
- Split the markdown contract: approval-markdown.ts is now the lightweight
  type + default const, and the stripMarkdown-backed downgrade moves to
  approval-markdown.runtime.ts. The channel registry (approvals.ts, on hot
  gateway/config startup) no longer pulls the markdown parser onto the hot path.
- Tighten the conformance scan to ignore comment lines so a declaration named
  only in prose cannot false-pass.
- Document the plaintext default as an intentional opt-in trade: undeclared
  third-party markdown channels downgrade on upgrade, deliberately, because
  plaintext can lose rendering but never leaks literal markers to an approver.
