+++ src/plugin-sdk/approval-reply-runtime.ts
@@ -17,11 +17,8 @@ export {
   type ExecApprovalReplyDecision,
   type ExecApprovalReplyMetadata,
 } from "../infra/exec-approval-reply.js";
-export {
-  downgradeApprovalMarkdownToPlaintext,
-  DEFAULT_APPROVAL_TEXT_MODE,
-  type ChannelApprovalTextMode,
-} from "./approval-markdown.js";
+export { DEFAULT_APPROVAL_TEXT_MODE, type ChannelApprovalTextMode } from "./approval-markdown.js";
+export { downgradeApprovalMarkdownToPlaintext } from "./approval-markdown.runtime.js";
 export { resolveExecApprovalCommandDisplay } from "../infra/exec-approval-command-display.js";
 export {
   resolveExecApprovalAllowedDecisions,
